Version 0.8.9
* Fixed handling of bigger files
* Fixed hash case sensitivity during cracking
* Improved several database structures (see hashtopus.sql)
+ Added new hashes and hashcat release

+++ server.php
@@ -348,7 +348,7 @@ switch ($action) {
     if (mysqli_num_rows($kver)==1) {
       $erej=mysqli_fetch_array($kver,MYSQLI_ASSOC);
       $chunktime=$erej["chunktime"];
-      $bench=intval($erej["benchmark"]);
+      $bench=floatval($erej["benchmark"]);
       $agid=$erej["id"];
       if ($erej["keyspace"]>0) {
         // we know the keyspace already
